requirements_cuda.txt
numpy
scipy
matplotlib
psutil

# https://arrayfire.com/
arrayfire

# https://docs.cupy.dev/en/stable/index.html
cupy-cuda12x

# https://numba.readthedocs.io/en/stable/
# https://nvidia.github.io/numba-cuda/
# replace numpy 2.4.1 -> 2.3.5 
numba-cuda[cu12]

# https://nvidia.github.io/cuda-python/latest/
cuda-python

# https://pypi.org/project/pycuda/
pycuda

# https://nvidia.github.io/warp/
warp-lang

# Only Linux
# https://triton-lang.org/main/getting-started/installation.html
# triton


# https://www.intel.com/content/www/us/en/developer/tools/oneapi/base-toolkit.html
intel-cmplr-lib-rt

# win32api
pywin32
